Added an new frequency option for "every 4 weeks"
KFP community meeting needs this option to schedule meeting every 4 weeks. This change is the easiest way to support this. Perhaps some time in the future we can open up RRULE to calendar.yaml file.

Tested locally by syncing the calendar. The calendar.yaml changes will be in a different commit.