Update the kubeflow-steering-committee-election-proposal.md (#664)
Made the following changes: 

Updated tool we will be using.  CIVS is what Elekto is based on and easier for people to administer. It's also what many of the CNCF projects use as well. 

Updated the timeline so that we account for the Industry Events, the holiday season and vacation time during this time of the year.