KSC Election Process Proposal.md (#665)
* Update kubeflow-steering-committee-election-proposal.md

Made the following changes: 

Updated tool we will be using.  CIVS is what Elekto is based on and easier for people to administer. It's also what many of the CNCF projects use as well. 

Updated the timeline so that we account for the Industry Events, the holiday season and vacation time during this time of the year.

* Update kubeflow-steering-committee-election-proposal.md

Added CIVS link and exception form statement and links to voter and candidate sections.

* Create elections.md

Added elections directory

* Delete elections.md

* Create guidelines-timeline-phases.md

created elections directory, kubeflow steering committee elections folder and guidelines-timeline-phases file

* Delete elections/kubeflow-steering-committee-elections-2023 directory

* Create kubeflow-steering-committee-elections-2023.md