Update kubeflow-steering-committee-elections-2023.md (#666)
* Update kubeflow-steering-committee-elections-2023.md

added Guidelines, timelines and phases information to the elections folder.

* Update kubeflow-steering-committee-elections-2023.md

Changed PSC to KSC

Added information to publishing the qualified voters and nominees list. 

Fixed the sentence about filing an exception.

* Update kubeflow-steering-committee-elections-2023.md

Added the following - *Please note The AOC is a suggested new committee and is here as a placeholder for a yet to be determined group that will work on non-code related related topics such marketing, communications, events and administrative needs of the project.